days: refactor DayAnimated & Item to separate components
