+++ src/MessageContainer/components/DayAnimated/styles.ts
@@ -1,0 +1,17 @@
+import { StyleSheet } from 'react-native'
+import Color from '../../../Color'
+
+export default StyleSheet.create({
+  dayAnimated: {
+    position: 'absolute',
+    width: '100%',
+  },
+  dayAnimatedDayContainerStyle: {
+    marginTop: 0,
+    marginBottom: 0,
+  },
+  dayAnimatedText: {
+    fontSize: 12,
+    color: Color.white,
+  },
+})
